tool/barcode/requirements.txt
elaphe==0.6.0
Pillow==2.3.0
